requirements/dev.txt
# Local development dependencies go here
-r base.txt

# Coverage
# ------------------------------------------------
coverage==4.2
django-coverage-plugin==1.3.1

# Utils
# ------------------------------------------------
django-extensions==1.7.4

# Debug
# ------------------------------------------------
django-debug-toolbar==1.6
ipdb==0.10.1

# Tests
# ------------------------------------------------
freezegun==0.3.11
pytest-django==3.4.3
pytest==3.10.0
